chore: trpc init moved to separate package to be reusable, packages typescript version cleanup
