+++ apps/tanstack-trpc-app/package.json
@@ -11,9 +11,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "@workspace/ui": "workspace:*",
-    "@workspace/tanstack-router": "workspace:*",
-    "@workspace/navigation": "workspace:*",
+    "@tailwindcss/vite": "^4.0.6",
     "@tanstack/react-query": "^5.66.5",
     "@tanstack/react-query-devtools": "^5.66.5",
     "@tanstack/react-router": "^1.114.3",
@@ -21,9 +19,12 @@
     "@tanstack/react-router-with-query": "^1.114.3",
     "@tanstack/react-start": "^1.114.3",
     "@trpc/client": "^11.0.0",
-    "@trpc/server": "^11.0.0",
-    "@trpc/tanstack-react-query": "^11.0.0",
-    "@tailwindcss/vite": "^4.0.6",
+    "@trpc/server": "^11.1.1",
+    "@trpc/tanstack-react-query": "^11.1.1",
+    "@workspace/navigation": "workspace:*",
+    "@workspace/tanstack-router": "workspace:*",
+    "@workspace/trpc": "workspace:*",
+    "@workspace/ui": "workspace:*",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "superjson": "^2.2.2",
@@ -32,10 +33,9 @@
     "zod": "^3.24.2"
   },
   "devDependencies": {
-    "@workspace/eslint-config": "workspace:*",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.10",
-    "typescript": "^5.7.2",
+    "@workspace/eslint-config": "workspace:*",    
     "vite": "^6.1.0"
   }
 }
