+++ apps/tanstack-trpc-app/src/components/TopMenu.tsx
@@ -10,15 +10,7 @@ const ITEMS = [
   {
     title: 'TanStack Query',
     to: '/demo/tanstack-query',
-  },
-  {
-    title: 'Start - Server functions',
-    to: '/demo/start/server-funcs',
-  },
-  {
-    title: 'Start - API Request',
-    to: '/demo/start/api-request',
-  },
+  }
 ] as AppTopMenuItem[];
 
 export const TopMenu = () => {
