+++ package.json
@@ -25,7 +25,7 @@
     "@workspace/eslint-config": "workspace:*",
     "nx": "20.5.0",
     "prettier": "^2.6.2",
-    "typescript": "~5.7.3",
+    "typescript": "^5.8.2",
     "verdaccio": "^6.0.5",
     "@testing-library/dom": "10.4.0",
     "@testing-library/react": "16.1.0",
