+++ packages/trpc/README.md
@@ -1,0 +1,11 @@
+# trpc
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build trpc` to build the library.
+
+## Running unit tests
+
+Run `nx test utils` to execute the unit tests via [Vitest](https://vitest.dev/).
