+++ packages/trpc/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@workspace/trpc",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./*": "./src/*.ts"
+  },
+  "devDependencies": {
+    "@types/node": "^22.13.10",
+    "@workspace/eslint-config": "workspace:*",
+    "@workspace/typescript-config": "workspace:*",
+    "jsonc-eslint-parser": "^2.4.0",
+    "vite": "^6.2.0",
+    "vite-plugin-dts": "^4.5.3"
+  },
+  "dependencies": {
+    "@trpc/server": "^11.1.1",
+    "superjson": "^2.2.2"
+  }
+}
